refactor : 새 프로젝트 구조에 맞추어 코드 수정 및 삭제
